backend/alembic/env.py
# backend/alembic/env.py
import os
import sys
from logging.config import fileConfig
from dotenv import load_dotenv

from alembic import context
from sqlalchemy import engine_from_config, pool

# Load environment variables from .env file
load_dotenv()

# Add app package path
sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))

# This is the Alembic Config object
config = context.config

# Interpret the config file for Python logging (commented out to avoid KeyError)
# if config.config_file_name is not None:
#     fileConfig(config.config_file_name)

# Import target metadata from app.models
from app import models  # noqa
target_metadata = models.Base.metadata

def get_url():
    return os.environ.get("DATABASE_URL")

def run_migrations_offline():
    url = get_url()
    context.configure(url=url, target_metadata=target_metadata, literal_binds=True)

    with context.begin_transaction():
        context.run_migrations()

def run_migrations_online():
    configuration = config.get_section(config.config_ini_section)
    configuration['sqlalchemy.url'] = get_url()
    connectable = engine_from_config(
        configuration,
        prefix='sqlalchemy.',
        poolclass=pool.NullPool
    )

    with connectable.connect() as connection:
        context.configure(connection=connection, target_metadata=target_metadata)

        with context.begin_transaction():
            context.run_migrations()

if context.is_offline_mode():
    run_migrations_offline()
else:
    run_migrations_online()
