backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
boto3
watchtower
python-dotenv
bcrypt==4.1.3
